Extract transport factory
