+++ src/main/java/com/amannmalik/mcp/api/McpServer.java
@@ -1,21 +1,18 @@
 package com.amannmalik.mcp.api;
 
-import com.amannmalik.mcp.auth.*;
 import com.amannmalik.mcp.codec.*;
 import com.amannmalik.mcp.core.*;
 import com.amannmalik.mcp.jsonrpc.*;
 import com.amannmalik.mcp.prompts.*;
 import com.amannmalik.mcp.roots.RootsManager;
 import com.amannmalik.mcp.spi.*;
-import com.amannmalik.mcp.transport.StdioTransport;
-import com.amannmalik.mcp.transport.StreamableHttpServerTransport;
+import com.amannmalik.mcp.transport.TransportFactory;
 import com.amannmalik.mcp.util.*;
 import jakarta.json.*;
 import jakarta.json.stream.JsonParsingException;
 
 import java.io.EOFException;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
@@ -84,7 +81,7 @@ public final class McpServer extends JsonRpcEndpoint implements AutoCloseable {
                      // It can be thought of like a "hint" to the model.
                      // For example, this information MAY be added to the system prompt.
                      String instructions) throws Exception {
-        super(createTransport(config),
+        super(TransportFactory.server(config),
                 new ProgressManager(new RateLimiter(config.progressPerSecond(),
                         config.rateLimiterWindowMs())),
                 config.initialRequestId());
@@ -143,38 +140,6 @@ public final class McpServer extends JsonRpcEndpoint implements AutoCloseable {
         }
         caps.add(ServerCapability.LOGGING);
         return EnumSet.copyOf(caps);
-    }
-
-    private static Transport createTransport(McpServerConfiguration config) throws Exception {
-        return switch (config.transportType()) {
-            case "stdio" -> new StdioTransport(System.in, System.out, config.defaultTimeoutMs());
-            case "http" -> {
-                if (!config.insecure() && config.authServers().isEmpty()) {
-                    throw new IllegalArgumentException("auth server must be specified");
-                }
-                AuthorizationManager authManager = null;
-                if (config.expectedAudience() != null && !config.expectedAudience().isBlank()) {
-                    var secretEnv = System.getenv("MCP_JWT_SECRET");
-                    var tokenValidator = secretEnv == null || secretEnv.isBlank()
-                            ? new JwtTokenValidator(config.expectedAudience())
-                            : new JwtTokenValidator(config.expectedAudience(), secretEnv.getBytes(StandardCharsets.UTF_8));
-                    authManager = new AuthorizationManager(List.of(new BearerTokenAuthorizationStrategy(tokenValidator)));
-                }
-                var ht = new StreamableHttpServerTransport(
-                        config,
-                        authManager);
-                if (config.verbose()) {
-                    if (config.serverPort() > 0) {
-                        System.err.println("Listening on http://127.0.0.1:" + ht.port());
-                    }
-                    if (config.httpsPort() > 0) {
-                        System.err.println("Listening on https://127.0.0.1:" + ht.httpsPort());
-                    }
-                }
-                yield ht;
-            }
-            default -> throw new IllegalArgumentException("Unknown transport type: " + config.transportType());
-        };
     }
 
     private ToolCallHandler createToolHandler(ToolProvider tools,
